Refactor: #37 - 글쓰기 오답노트 섹션 분리
