+++ src/pages/wrong-notes/index.tsx
@@ -5,9 +5,7 @@ import PageHeader from "@/components/PageHeader";
 import Button from "@/components/ui/Button";
 import Chip from "@/components/ui/Chip";
 import SelectBox from "@/components/ui/SelectBox";
-import ToggleButtonGroup from "@/components/ui/ToggleButtonGroup";
 import {
-  saveWrongNote,
   getWrongNotes,
   deleteWrongNote,
   getFriendsSharedWrongNotes,
@@ -15,18 +13,8 @@ import {
   type WrongNote,
 } from "@/firebase/services";
 import type { FriendInfo } from "@/types/friendship.types";
+import { wrongNoteSortOptions } from "@/constants/options.constants";
 import {
-  categoryOptions,
-  languageOptions,
-  platformOptions,
-  programmersGrades,
-  resultOptions,
-  tagOptions,
-  wrongNoteSortOptions,
-} from "@/constants/options.constants";
-import type { Filters, FormData } from "@/types/wrong-notes.types";
-import {
-  baekjoonGrades,
   getCategoryLabel,
   getGradeLabel,
   getLanguageLabel,
@@ -34,40 +22,27 @@ import {
   getResultLabel,
   getTagLabels,
 } from "@/utils/options.utils";
-import CodeEditorGroup from "@/components/CodeEditorGroup";
 import WrongNoteFilter from "@/components/WrongNoteFilter";
+import { useWrongNoteFilter } from "@/hooks/wront-notes/useWrongNoteFilter";
+import { WriteSection } from "@/components/wrong-notes/write";
 
 export default function WrongNotes() {
   const [activeTab, setActiveTab] = useState<"write" | "list" | "friends">(
     "list",
   );
-  const [formData, setFormData] = useState<FormData>({
-    title: "",
-    link: "",
-    language: "",
-    date: new Date().toISOString().split("T")[0],
-    platform: "",
-    category: "",
-    grade: "",
-    myCode: "",
-    solution: "",
-    comment: "",
-    share: false,
-    tags: [],
-    result: "",
-  });
-  const [isSubmitting, setIsSubmitting] = useState(false);
   const [notes, setNotes] = useState<WrongNote[]>([]);
   const [isLoading, setIsLoading] = useState(false);
-  const [sortBy, setSortBy] = useState("latest");
-  const [filters, setFilters] = useState<Filters>({
-    platform: "",
-    category: "",
-    language: "",
-    result: "",
-    tag: "",
-  });
-  const [searchQuery, setSearchQuery] = useState("");
+
+  const {
+    filters,
+    sortBy,
+    searchQuery,
+    setSearchQuery,
+    handleFilterChange,
+    handleSortByChange,
+    clearFilters,
+    hasActiveFilters,
+  } = useWrongNoteFilter();
   const navigate = useNavigate();
 
   // 친구 오답노트 관련 state
@@ -104,7 +79,7 @@ export default function WrongNotes() {
 
         return 0;
       });
-  }, [friendNotes, filters, searchQuery, sortBy]);
+  }, [friendNotes, filters, friendFilter, searchQuery, sortBy]);
 
   // 친구 이름 가져오기 helper
   const getFriendDisplayName = (userId: string): string => {
@@ -144,27 +119,6 @@ export default function WrongNotes() {
       });
   }, [notes, filters, searchQuery, sortBy]);
 
-  const handleFilterChange = (key: keyof Filters, value: string) => {
-    setFilters((prev) => ({ ...prev, [key]: value }));
-  };
-
-  const clearFilters = () => {
-    setFilters({
-      platform: "",
-      category: "",
-      language: "",
-      result: "",
-      tag: "",
-    });
-    setSearchQuery("");
-  };
-
-  const hasActiveFilters = useMemo(() => {
-    return (
-      !!filters.platform || !!filters.result || !!filters.tag || !!searchQuery
-    );
-  }, [filters.platform, filters.result, filters.tag, searchQuery]);
-
   // 목록 불러오기
   const loadNotes = async () => {
     try {
@@ -201,63 +155,8 @@ export default function WrongNotes() {
     } else if (activeTab === "friends") {
       loadFriendNotes();
     }
+    clearFilters();
   }, [activeTab]);
-
-  const handleInputChange = (
-    field: keyof FormData,
-    value: string | boolean | string[],
-  ) => {
-    setFormData((prev) => ({ ...prev, [field]: value }));
-  };
-
-  const handlePlatformChange = (value: string) => {
-    setFormData((prev) => ({ ...prev, platform: value, grade: "" }));
-  };
-
-  const handleSortByChange = (value: string) => {
-    setSortBy(value);
-  };
-
-  const getGradeOptions = () => {
-    if (formData.platform === "programmers") return programmersGrades;
-    if (formData.platform === "baekjoon") return baekjoonGrades;
-    return [];
-  };
-
-  const resetForm = () => {
-    setFormData({
-      title: "",
-      link: "",
-      language: "",
-      date: new Date().toISOString().split("T")[0],
-      platform: "",
-      grade: "",
-      category: "",
-      myCode: "",
-      solution: "",
-      comment: "",
-      share: false,
-      tags: [],
-      result: "",
-    });
-  };
-
-  const handleSubmit = async () => {
-    if (isSubmitting) return;
-
-    try {
-      setIsSubmitting(true);
-      await saveWrongNote(formData);
-      alert("오답노트가 저장되었습니다.");
-      resetForm();
-      setActiveTab("list");
-    } catch (error) {
-      console.error("저장 실패:", error);
-      alert(error instanceof Error ? error.message : "저장에 실패했습니다.");
-    } finally {
-      setIsSubmitting(false);
-    }
-  };
 
   const handleDelete = async (noteId: string) => {
     if (!confirm("정말 삭제하시겠습니까?")) return;
@@ -269,6 +168,10 @@ export default function WrongNotes() {
       console.error("삭제 실패:", error);
       alert("삭제에 실패했습니다.");
     }
+  };
+
+  const handleActiveTab = () => {
+    setActiveTab("list");
   };
 
   return (
@@ -596,189 +499,7 @@ export default function WrongNotes() {
         )}
 
         {/* 작성 탭 */}
-        {activeTab === "write" && (
-          <div className="mt-6 space-y-6">
-            {/* 문제 이름 & 언어 */}
-            <div className="grid grid-cols-1 md:grid-cols-[1fr_200px] gap-4">
-              <div>
-                <label className="block text-sm font-medium text-text mb-2">
-                  문제 이름
-                </label>
-                <input
-                  type="text"
-                  value={formData.title}
-                  onChange={(e) => handleInputChange("title", e.target.value)}
-                  placeholder="예: 두 수의 합, 타겟 넘버 등"
-                  className="w-full px-4 py-2 text-sm outline outline-1 outline-border rounded-md bg-surface text-text
-                    hover:outline-primary focus:outline-primary focus:ring-2 focus:ring-blue-200 transition-all"
-                />
-              </div>
-              <div>
-                <label className="block text-sm font-medium text-text mb-2">
-                  언어
-                </label>
-                <SelectBox
-                  options={languageOptions}
-                  value={formData.language}
-                  onChange={(e) =>
-                    handleInputChange("language", e.target.value)
-                  }
-                  placeholder="언어 선택"
-                  fullWidth
-                />
-              </div>
-            </div>
-
-            {/* 문제 링크 */}
-            <div>
-              <label className="block text-sm font-medium text-text mb-2">
-                문제 링크
-              </label>
-              <input
-                type="url"
-                value={formData.link}
-                onChange={(e) => handleInputChange("link", e.target.value)}
-                placeholder="https://programmers.co.kr/..."
-                className="w-full px-4 py-2 text-sm outline outline-1 outline-border rounded-md bg-surface text-text
-                  hover:outline-primary focus:outline-primary focus:ring-2 focus:ring-blue-200 transition-all"
-              />
-            </div>
-
-            {/* 날짜 & 플랫폼 & 등급 */}
-            <div className="grid grid-cols-1 md:grid-cols-4 gap-4">
-              <div>
-                <label className="block text-sm font-medium text-text mb-2">
-                  날짜
-                </label>
-                <input
-                  type="date"
-                  value={formData.date}
-                  onChange={(e) => handleInputChange("date", e.target.value)}
-                  className="w-full px-4 py-2 text-sm outline outline-1 outline-border rounded-md bg-surface text-text
-                    hover:outline-primary focus:outline-primary focus:ring-2 focus:ring-blue-200 transition-all"
-                />
-              </div>
-              <div>
-                <label className="block text-sm font-medium text-text mb-2">
-                  알고리즘
-                </label>
-                <SelectBox
-                  options={categoryOptions}
-                  value={formData.category}
-                  onChange={(e) =>
-                    handleInputChange("category", e.target.value)
-                  }
-                  placeholder="알고리즘 선택"
-                  fullWidth
-                />
-              </div>
-              <div>
-                <label className="block text-sm font-medium text-text mb-2">
-                  플랫폼
-                </label>
-                <SelectBox
-                  options={platformOptions}
-                  value={formData.platform}
-                  onChange={(e) => handlePlatformChange(e.target.value)}
-                  placeholder="플랫폼 선택"
-                  fullWidth
-                />
-              </div>
-              <div>
-                <label className="block text-sm font-medium text-text mb-2">
-                  등급
-                </label>
-                <SelectBox
-                  options={getGradeOptions()}
-                  value={formData.grade}
-                  onChange={(e) => handleInputChange("grade", e.target.value)}
-                  placeholder={
-                    formData.platform ? "등급 선택" : "플랫폼을 먼저 선택"
-                  }
-                  disabled={!formData.platform}
-                  fullWidth
-                />
-              </div>
-            </div>
-
-            {/* 제출 결과 & 작성 이유 */}
-            <div className="flex w-full justify-between items-start">
-              <div className="w-[50%]">
-                <label className="block text-sm font-medium text-text mb-2">
-                  제출 결과
-                </label>
-                <ToggleButtonGroup
-                  options={resultOptions}
-                  value={formData.result}
-                  onChange={(value) => handleInputChange("result", value)}
-                />
-              </div>
-              <div className="w-[50%]">
-                <label className="block text-sm font-medium text-text mb-2">
-                  작성 이유 (복수 선택 가능)
-                </label>
-                <ToggleButtonGroup
-                  options={tagOptions}
-                  value={formData.tags}
-                  onChange={(value) => handleInputChange("tags", value)}
-                  multiple
-                />
-              </div>
-            </div>
-            <CodeEditorGroup
-              isEditMode={true}
-              language={formData.language}
-              myCode={formData.myCode}
-              solution={formData.solution}
-              onChangeMyCode={(value) => handleInputChange("myCode", value)}
-              onChangeSolution={(value) => handleInputChange("solution", value)}
-            />
-
-            {/* 코멘트 */}
-            <div>
-              <label className="block text-sm font-medium text-text mb-2">
-                코멘트
-              </label>
-              <textarea
-                value={formData.comment}
-                onChange={(e) => handleInputChange("comment", e.target.value)}
-                placeholder="이 문제에서 배운 점, 주의할 점 등을 기록하세요..."
-                rows={4}
-                className="w-full px-4 py-3 text-sm outline outline-1 outline-border rounded-md bg-surface text-text
-                  hover:outline-primary focus:outline-primary focus:ring-2 focus:ring-blue-200 transition-all resize-none"
-              />
-            </div>
-
-            {/* 공유하기 */}
-            <div className="flex items-center gap-3">
-              <input
-                type="checkbox"
-                id="share"
-                checked={formData.share}
-                onChange={(e) => handleInputChange("share", e.target.checked)}
-                className="w-4 h-4 rounded border-border text-primary focus:ring-primary cursor-pointer"
-              />
-              <label
-                htmlFor="share"
-                className="text-sm text-text cursor-pointer"
-              >
-                다른 사용자에게 공유하기
-              </label>
-            </div>
-
-            {/* 제출 버튼 */}
-            <div className="flex justify-end pt-4">
-              <Button
-                variant="primary"
-                size="lg"
-                onClick={handleSubmit}
-                disabled={isSubmitting}
-              >
-                {isSubmitting ? "저장 중..." : "오답노트 저장"}
-              </Button>
-            </div>
-          </div>
-        )}
+        {activeTab === "write" && <WriteSection cb={handleActiveTab} />}
       </div>
     </div>
   );
